Develop to Master

+++ src/main/java/com/onulstore/service/MemberService.java
@@ -1,9 +1,9 @@
 package com.onulstore.service;
 
 import com.onulstore.config.SecurityUtil;
-import com.onulstore.config.exception.Exception;
+import com.onulstore.config.exception.CustomException;
 import com.onulstore.domain.enums.Authority;
-import com.onulstore.domain.enums.ErrorResult;
+import com.onulstore.domain.enums.CustomErrorResult;
 import com.onulstore.domain.member.Member;
 import com.onulstore.domain.member.MemberRepository;
 import com.onulstore.domain.product.Product;
@@ -13,11 +13,13 @@ import com.onulstore.web.dto.ProductDto;
 import com.onulstore.web.dto.ProductDto.ProductResponse;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,31 +36,43 @@ public class MemberService {
     public MemberDto.MemberResponse getMyInfo() {
         return memberRepository.findById(SecurityUtil.getCurrentMemberId())
             .map(MemberDto.MemberResponse::of)
-            .orElseThrow(() -> new Exception(ErrorResult.NOT_EXIST_USER));
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
     }
 
     @Transactional
     public MemberDto.MemberResponse updateProfile(MemberDto.UpdateRequest updateRequest) {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
         Member updateMember = member.updateProfile(updateRequest);
         return MemberDto.MemberResponse.of(memberRepository.save(updateMember));
     }
 
     @Transactional
     public void deleteProfile() {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
         memberRepository.delete(member);
     }
 
     @Transactional
     public void updatePassword(PasswordDto passwordDto) {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
 
         if (!passwordDto.getNewPassword().equals(passwordDto.getNewPasswordConfirm())) {
-            throw new Exception(ErrorResult.UPDATE_PASSWORD);
+            throw new CustomException(CustomErrorResult.UPDATE_PASSWORD);
         }
 
         Member updateMember = member.updatePassword(
@@ -68,8 +82,12 @@ public class MemberService {
 
     @Transactional
     public ArrayList<ProductResponse> latestProduct(HttpServletRequest request) {
-        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-            () -> new Exception(ErrorResult.NOT_EXIST_USER));
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
         HttpSession session = request.getSession();
 
         ArrayList<Product> latest = (ArrayList) session.getAttribute("List");
@@ -81,12 +99,20 @@ public class MemberService {
     }
 
     @Transactional
-    public List<Integer> memberDashBoard(LocalDateTime localDateTime){
-        List<Member> memberList = memberRepository.findAllByAuthorityAndCreatedDateAfter(Authority.ROLE_USER, localDateTime);
-        List<Member> sellerList = memberRepository.findAllByAuthorityAndCreatedDateAfter(Authority.ROLE_SELLER, localDateTime);
-        List<Integer> memberAmount = new ArrayList<>();
-        memberAmount.add(memberList.size());
-        memberAmount.add(sellerList.size());
+    public List<Long> memberDashBoard(LocalDateTime localDateTime){
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal()
+            .equals("anonymousUser")) {
+            throw new CustomException(CustomErrorResult.LOGIN_NEEDED);
+        }
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId())
+            .orElseThrow(() -> new CustomException(CustomErrorResult.NOT_EXIST_USER));
+        if (!member.getAuthority().equals(Authority.ROLE_ADMIN.getKey())) {
+            throw new CustomException(CustomErrorResult.ACCESS_PRIVILEGE);
+        }
+
+        Long members = memberRepository.countByAuthorityAndCreatedDateAfter(Authority.ROLE_USER, localDateTime);
+        Long sellers = memberRepository.countByAuthorityAndCreatedDateAfter(Authority.ROLE_SELLER, localDateTime);
+        List<Long> memberAmount = Arrays.asList(members, sellers);
         return memberAmount;
     }
 }
